PyTorch/StreetScene/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
tqdm>=4.60.0
pyyaml>=5.4.0
tensorboard>=2.5.0
ultralytics>=8.0.0
scikit-learn>=0.24.0
motmetrics>=1.2.0